+++ mmdet/models/dense_heads/dense_test_mixins.py
@@ -34,12 +34,8 @@ class BBoxTestMixin(object):
                 with shape (n,)
         """
         outs = self.forward(feats)
-        if getattr(self, 'static', False):
-            results_list = self.get_bboxes_statically(
-                *outs, img_metas=img_metas, rescale=rescale)
-        else:
-            results_list = self.get_bboxes(
-                *outs, img_metas=img_metas, rescale=rescale)
+        results_list = self.get_bboxes(
+            *outs, img_metas=img_metas, rescale=rescale)
         return results_list
 
     def aug_test_bboxes(self, feats, img_metas, rescale=False):
